优化 Pag.devI 生成模版过期判定

+++ absir-developer/src/main/java/com/absir/aserv/support/developer/IDeveloper.java
@@ -23,9 +23,9 @@ public interface IDeveloper {
 
     public int getDeveloperNewType();
 
-    public long getDeveloperNewTime();
+    public int getDeveloperNewTime();
 
-    public void setDeveloperNewTime(long developerNewTime);
+    public void setDeveloperNewTime(int developerNewTime);
 
     public IModel getModelEntity(JoEntity joEntity);
