+++ absir-developer/src/main/java/com/absir/aserv/support/developer/IRender.java
@@ -18,7 +18,7 @@ public interface IRender {
 
     public static final IRender ME = BeanFactoryUtils.get(IRender.class);
 
-    public String dev(long devTime);
+    public String dev(int devTime);
 
     public String echo(String value);
